+++ src/routes/+page.svelte
@@ -2,16 +2,33 @@
    import "../app.css";
 
    import { Stage } from "$lib/index.js";
-   import Title from './_slides/Title.svelte'
+   import Title from "./_slides/Title.svelte";
+   import BabysFirstSlide from "./_slides/BabysFirstSlide.svelte";
 
-   import { PresentationViewer } from '$lib/index.js'
+   import { PresentationViewer } from "$lib/index.js";
+   import HighlightProvider from "./_components/HighlightProvider.svelte";
+
+   let lastRefresh = $state(0);
+
+   $effect(() => {
+      if (import.meta.hot) {
+         lastRefresh = Date.now();
+      }
+   });
+
 </script>
 
 <div>
-   <Stage>
-      <Title />
-   </Stage>
+   {#key lastRefresh}
+      <Stage>
+         <HighlightProvider>
+            <Title />
+            <BabysFirstSlide />
+         </HighlightProvider>
+      </Stage>
+   {/key}
 </div>
+
 <!-- <PresentationViewer /> -->
 
 <style>