+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Built very basic website
 - Added MIT License
 
+- implemented container units for stage, tied to `cqmin`
+- added reactive properties for internal slide state management
+
 ## [0.0.2] - 2025-08-26
 
 - Initial public alpha release! 🎉